ENH: Changed `bit_count` UFunc return to int8
WIP: Fix the typing

+++ numpy/core/code_generators/ufunc_docstrings.py
@@ -4245,6 +4245,7 @@ add_newdoc('numpy.core.umath', 'bit_count',
     10
     >>> a = np.array([2**i - 1 for i in range(16)])
     >>> np.bit_count(a)
-    array([ 0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15])
+    array([ 0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15],
+          dtype=uint8)
 
     """)
